+++ apps/reports/views/course.swig
@@ -1,4 +1,4 @@
-{% extends '../admin.swig' %}
+{% extends '../../../src/views/admin.swig' %}
 
 {% block heading %}Course Report: {{ course }}{% endblock %}
