+++ apps/reports/views/partials/items-list.swig
@@ -16,15 +16,15 @@
 					<tr>
 						<td>
 							{% set status = item.status %}
-							{% include '../../items/partials/status.swig' %}
+							{% include '../../../items/views/partials/status.swig' %}
 						</td>
 						<td><span data-clipboard-text="{{ item.barcode }}" class="clipboard">{{ item.barcode }}</span></td>
 						<td><a href="/items/{{ item._id | raw }}">{{ item.name }}</a></td>
 						<td>{% if item.owner %}<a href="/users/{{ item.owner._id | raw }}">{{ item.owner.name }}</a>{% else %}&mdash;{% endif %}</td>
 						{% set department = item.department %}
-						<td>{% include '../../departments/partials/department.swig' %}</td>
+						<td>{% include '../../../departments/views/partials/department.swig' %}</td>
 						{% set group = item.group %}
-						<td>{% include '../../groups/partials/group.swig' %}</td>
+						<td>{% include '../../../groups/views/partials/group.swig' %}</td>
 					</tr>
 				{% endfor %}
 			</tbody>
